# perf(es/minifier): Pre-allocate in analyzer